Realized that perturbations per perturbation are different due to dosage level
A unique perturbation should be based on perturbation and drug dosage, to work with scButterfly. Currently we have been treat all dosages as one perturbation, and the model wasn't performing well. On the other hand the number of stimulated cells per perturbation and drug dosage is very low. Are there other datasets?