Compare scpregan and scpregan reproducible
Checking theirr benchmarking in the paper of scpregan, I get overall better results for both versions of scpregan. I will stick with the one provided by the main repo, not the reproducible, due to resources, it takes much more time for the second one to finish. Also this is indication of the variance of the performance of the models. That's why I should attempt to run it more times with error bars