Add vidr for single and mutliple doses
The dataset pipelines need to be redesigned and find a common approach. Should the datasets that don't have a concept of drug to assume as Dose 0, what should be the common keys for dosage, condition, single, and multiple dosages, perturbations? We need an abstraction of splitting to train, test, control, perturb data