Experimenting with adv loss coeff
Our adversarial loss is computed by swapping the labels for the control and perturbed samples.
The adversarial loss doesn't seem to help. The lower the factor of the adversarial loss,
the better the results. Although the no adversarial run wasn't as good as the low coeff adv.
But it is worth mentioning that DEGs have much higher variability compared to r2 metrics.

For benchmarking we should repeat it for 3 different seeds, and check the bar errors and the variability.

Next step is to try to treat control samples as real and fake the perturbed. The goal would be the perturb to
be aligned with the perturb, aka the encoder to learn perturbation free information.

I did an ablation study of butterfly as well. Raised a github issue. The translator loss, rather than the discriminator

I have separated the training to a separate class. We can delay adv loss with the `discr_good_enough_epoch_threshold` as well.

is the significant factor.